chore: Update to latest dotnet framework
This latest framework is multi platform so that the lib can be used in
more situations.

Removed unused Microsoft.CSharp and System.Text.Encoding.CodePages references
which are not provided anymore in latests dotnet framework.

Silenced a warning on BinaryFormatter about security issues. As this is about
deserializing objects from unknown sources, which is not the case here,
it is supposed to be safe. Should deal with this later on just for the
sake of it.

Fixed two re-throw situations which are also new warnings because it
changes the stack of the exception with the new throw location, losing
the original one.