feature: Add a basic app that uses the lib
The add takes a list of file on its cmdline and, for each file, convert
BIN files to CSV files, or patches BIN files from CSV files.

The BIN to CSV is a "full conversion", as it exports all valid columns
and all valid entries.

The CSV to BIN is a "patch": it requires the BIN file to also exists and
read data from CSV to overwrite entries. The CSV file thus does not
require to have all entries or columns. It requires to have a "ROW-ID"
and/or a "ROW-NAME" to match entries. If both are provided, the "ROW-ID"
is used.

Using this system, mods can be provided as CSV files containing only the
changes to make. So that it improve compatibility with other mods,
weight of the mod itself and locale support.